fixing existing tests
